refactor: extract rename safety checks
